Trait basilisk_runtime::Filter

pub trait Filter<T> {
    fn filter(&T) -> bool;
}

Simple trait for providing a filter over a reference to some type.

Required methods

fn filter(&T) -> bool

Determine if a given value should be allowed through the filter (returns true) or not.